Played around with some more jquery animations etc.
Created a userdashboard pages (currently does not have anything useful in it, but we will make it the user dashboard after user is authenticated - its location can be changed as well, in root atm)

Authentication now redirects to the user dashboard. User dashboard calls confirmLogin() to check for logged in user.