Added folders to the scripts path, to organize the scripts. Note that only the scripts in 'include' are to be included in other server-side php files. All other scripts are responsible for handling particular functionality e.g. registration, login, logout
These will be particularly useful for javascript/jquery AJAX calls later. We should restrict the php scripts to be server side - instead of 'including' any of them in the client side pages. If required, use javascript to GET/POST to server-side php scripts for any information that is required on a page.