HUGE commit.
Incorporated a dashboard library to work with our system. Added some basic project functions and tested these out with ajax. The dashboard allows widgets to be placed and moved around between the panels. Each widget has its html page - therefore it can run javascript, ajax etc.

Made an example one that lists all the projects lead by the currently-logged-in user. Can also add projects, which POSTs the data to the database via AJAX.

Can expand on this a lot and improve it to suit our needs :)