Added a test index and login script - STILL UNDER CONSTRUCTION SO DO NOT EDIT YET.
login2.js --> Basically transitioned the dialog login form into one that is embedded on the page (this will let us add hover event handlers and other stuff later). Wrote a bunch of functions for showing and hiding the login form, front-end validation and reporting errors like it did in the dialog form. Still some small bugs, but will work on that tomorrow to fix these.