Added some new functionality. The ability to log hours to an issue --> so a new table was created to map an issue to a 'hour log'. So a user can record hours worked on an issue multiple times. e.g. on day 1 they do 2 hours of research, they can log that as a issuehour record. Day 2, they do 5 hours of coding, they can log this to the same issue by submitting another work log. We can then make queries to automatically add up the hours of all the 'work logs' for each issue to get the total nubmer of hours worked on that issue. This functionality also lets users be able to see how many hours each user spent on each issue (which is important if the issue is a BIG issue). To extend it further we can store the date when a issuehour log is created - this way users can see when a project member did work on a particular issue. etc.
The latest database file has been updated: tracey.sql

The button for logging work hours is on the issue interface itself - but for now i've kept it out of sight --> you will have to look right at the bottom of the issue and click on the 'Log Hours' link. This will open up another small light box, which u fill in 2 fields and submit. You can do this multiple times within the same issue if you want as well.

/end of essay