Added missing functionality:
Modifying an issue will now update the lastmodificationdate. If the issue status gets changed to resolved (and it wasnt previously resolved) then a resolveddate will be set in the db.

If the user is creating an issue, the start date, last modification date and resolved date will not show up as fields (as they are not for input) - they will show up when viewing an existing issue.

The issuestatus drop down menu is now populated from values in the database, instead of hardcoded values.

The default priority has been changed to Normal.