Lots of additions. Added some new functions to project and issue include scripts that are handy for getting work log data. Project dashboard now has a new widget that shows recently submitted work logs - with the ability to view more detail about the work log by clicking the 'view log' button on the list.
The tables on project dashboard widgets are now sortable (will do the same for userdashboard later).

The table of issues on the project dashboard now contains the View button as well - to look the same as the table in user dashboard.

Creating a project will now successfully store the creation date.