LOTS of additions / changes.
Component feature added (along with back-end stuff I needed for it that was missing). Can create components for a project, modify them etc.

Users can also 'join' a component from the component widget in the project dashboard. This automatically links them to the component in the database.

Fixed the scrollbars showing up on lightboxes - looks neater.

Modified the datatable library so that it can integrate with the themeroller library. The tables look a fair bit nicer now!

The database file has been updated (tracey.sql) - Make sure you update to this. Components now store a due date as well as a foreign-key userid of the creator.

Created a new widget that lists issues  related to the components that a user is a 'member of'. This widget is not a default widget, and so must be added from the 'add widget' menu.

Fixed the problem where work logs could be submitted when no hours were specified. Work logs require number of hours to be specified before it is stored in the database now.

Performed some jQuery code optimization based on cheatsheets to reduce the console load time of widgets (as seen on Firebug's console.time tests).

Components can be viewed in the same way as issues, using the component interface and lightbox. The labels are slightly different depending on whether you are creating a component or viewing/modifying one. Modification of a component works as well i.e. changes can be saved when viewing one.